frontend: modularize Post component at thread page
