+++ frontend/src/pages/thread/[threadId].tsx
@@ -1,21 +1,6 @@
 import { BBSHeading, BBSHeadingTitle } from "@/components/BBSHeading";
 import BBSLayout from "@/components/BBSLayout";
-import {
-    Box,
-    Button,
-    chakra,
-    FormControl,
-    HStack,
-    Input,
-    Table,
-    TableContainer,
-    Tbody,
-    Td,
-    Text,
-    Tr,
-    useToast,
-    VStack,
-} from "@chakra-ui/react";
+import { Button, chakra, FormControl, HStack, Input, Table, TableContainer, Tbody, Td, Text, Tr, useToast, VStack } from "@chakra-ui/react";
 import Head from "next/head";
 import { getDeBBSAddress } from "@/constants/ContractAddresses";
 import { deBbsAbi } from "@/generated";
@@ -27,8 +12,7 @@ import { waitForTransactionReceipt, writeContract } from "wagmi/actions";
 import { formatEther } from "viem";
 import { convertTimestampToLocalTime } from "@/utils/convertTimestampToLocalTime";
 import { useRouter } from "next/router";
-import { EnsNameOrAddress } from "@/components/EnsNameOrAddress";
-import * as jdenticon from "jdenticon";
+import Post from "@/components/Post";
 
 export default function Thread() {
     const { chain } = useAccount();
@@ -158,27 +142,7 @@ export default function Thread() {
                     </TableContainer>
 
                     <Hr borderStyle="dashed" my={2} />
-                    {getPostsByThreadResult &&
-                        getPostsByThreadResult.map((post, i) => {
-                            const svgString = jdenticon.toSvg(post.postOwner, 20);
-                            return (
-                                <Box key={i}>
-                                    <VStack spacing={2} align="start">
-                                        <HStack justify="left">
-                                            <Box border={`1px solid ${primaryColor}`}>
-                                                <Box w={5} h={5} dangerouslySetInnerHTML={{ __html: svgString }}></Box>
-                                            </Box>
-                                            <Text>
-                                                <EnsNameOrAddress address={post.postOwner} shorten />
-                                            </Text>
-                                        </HStack>
-                                        <Text>{post.postContent}</Text>
-                                        <Text>[{convertTimestampToLocalTime(Number(post.timestamp))}]</Text>
-                                    </VStack>
-                                    <Hr borderStyle="dashed" my={2} />
-                                </Box>
-                            );
-                        })}
+                    {getPostsByThreadResult && getPostsByThreadResult.map((post, i) => <Post key={i} post={post} />)}
 
                     <BBSHeading headingProps={{ mt: 6, mb: 2 }}>&gt; Create A Thread</BBSHeading>
                     <FormControl as="form" onSubmit={handleSubmit}>
